server/requirements.txt
langchain==0.0.340
openai==1.3.4
PyYAML==6.0.1
tiktoken==0.7.0
Flask>=3.0.3
openai
stability-sdk
google-cloud-texttospeech